+++ ali/common/src/main/java/com/yanny/ali/manager/EnumTranslation.java
@@ -1,0 +1,4 @@
+package com.yanny.ali.manager;
+
+public record EnumTranslation(String modId, String owner) {
+}