+++ ali/common/src/main/java/com/yanny/ali/plugin/glm/GlobalLootModifierUtils.java
@@ -8,9 +8,6 @@ import com.yanny.ali.api.ILootModifier;
 import com.yanny.ali.api.IOperation;
 import com.yanny.ali.api.IServerUtils;
 import com.yanny.ali.plugin.common.nodes.GlobalLootModifierNode;
-import com.yanny.ali.plugin.mods.BaseAccessor;
-import com.yanny.ali.plugin.mods.ClassAccessor;
-import com.yanny.ali.plugin.mods.ReflectionUtils;
 import com.yanny.ali.plugin.server.TooltipUtils;
 import net.minecraft.advancements.predicates.entity.EntityPredicate;
 import net.minecraft.advancements.predicates.entity.EntityTypePredicate;
@@ -167,21 +164,6 @@ public class GlobalLootModifierUtils {
                 return c instanceof AnyOfCondition condition && tablePredicate(condition.terms, location, predicate);
             }
         });
-    }
-
-    public static <T extends BaseAccessor<?> & IGlobalLootModifierAccessor> void registerGlobalLootModifier(IGlobalLootModifierPlugin.IRegistry registry, Class<T> clazz, ILootTableIdConditionPredicate predicate) {
-        ClassAccessor classAnnotation = clazz.getAnnotation(ClassAccessor.class);
-
-        if (classAnnotation != null) {
-            try {
-                Class<?> functionClass = Class.forName(classAnnotation.value());
-                registry.registerGlobalLootModifier(functionClass, (u, c) -> ReflectionUtils.copyClassData(clazz, c).getLootModifier(u, predicate));
-            } catch (Throwable e) {
-                LOGGER.warn("Failed to register GLM for {} with error {}", classAnnotation.value(), e.getMessage(), e);
-            }
-        } else {
-            throw new IllegalStateException("Missing ClassAccessor annotation for GLM " + clazz.getName());
-        }
     }
 
     public static Optional<ILootModifier<?>> getMissingGlobalLootModifier(IServerUtils utils, IGlobalLootModifierWrapper modifier, ILootTableIdConditionPredicate predicate) {