+++ ali/common/src/main/java/com/yanny/ali/plugin/mods/BaseAccessor.java
@@ -1,9 +1,0 @@
-package com.yanny.ali.plugin.mods;
-
-public abstract class BaseAccessor<T> {
-    protected final T parent;
-
-    public BaseAccessor(T parent) {
-        this.parent = parent;
-    }
-}