+++ ali/common/src/main/java/com/yanny/ali/plugin/mods/IConsumeEffectTooltip.java
@@ -1,8 +1,0 @@
-package com.yanny.ali.plugin.mods;
-
-import com.yanny.aci.tooltip.TooltipBuilder;
-import com.yanny.ali.api.IServerUtils;
-
-public interface IConsumeEffectTooltip {
-    TooltipBuilder getTooltip(IServerUtils utils);
-}