+++ ali/common/src/main/java/com/yanny/ali/plugin/mods/IItemStackModifier.java
@@ -1,8 +1,0 @@
-package com.yanny.ali.plugin.mods;
-
-import com.yanny.ali.api.IServerUtils;
-import net.minecraft.world.item.ItemStack;
-
-public interface IItemStackModifier {
-    ItemStack applyItemStackModifier(IServerUtils utils, ItemStack itemStack);
-}