+++ ali/forge/src/main/java/com/yanny/ali/forge/plugin/mods/farmers_delight/package-info.java
@@ -1,4 +1,0 @@
-@ParametersAreNonnullByDefault
-package com.yanny.ali.forge.plugin.mods.farmers_delight;
-
-import javax.annotation.ParametersAreNonnullByDefault;