+++ alicompat/common/src/main/java/com/yanny/alicompat/accessor/IConditionTooltip.java
@@ -1,0 +1,8 @@
+package com.yanny.alicompat.accessor;
+
+import com.yanny.aci.tooltip.TooltipBuilder;
+import com.yanny.ali.api.IServerUtils;
+
+public interface IConditionTooltip {
+    TooltipBuilder getTooltip(IServerUtils utils);
+}