+++ alicompat/fabric/src/main/java/com/yanny/alicompat/fabric/platform/FabricCompatPlatform.java
@@ -1,0 +1,11 @@
+package com.yanny.alicompat.fabric.platform;
+
+import com.yanny.alicompat.platform.ICompatPlatform;
+import net.fabricmc.loader.api.FabricLoader;
+
+public class FabricCompatPlatform implements ICompatPlatform {
+    @Override
+    public boolean isModLoaded(String modId) {
+        return FabricLoader.getInstance().isModLoaded(modId);
+    }
+}