+++ alicompat/forge/src/main/java/com/yanny/alicompat/forge/datagen/package-info.java
@@ -1,0 +1,4 @@
+@ParametersAreNonnullByDefault
+package com.yanny.alicompat.forge.datagen;
+
+import javax.annotation.ParametersAreNonnullByDefault;